requirements_min.txt
# Versão mínima das dependências
dash
pandas
requests
python-dotenv
